% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available.R
\name{get_available}
\alias{get_available}
\title{Extract available units}
\usage{
get_available(
  filepath,
  section = c("off-site", "on-site"),
  module = c("area", "hedgerow", "watercourse")
)
}
\arguments{
\item{filepath}{The path to the metric spreadsheet.}

\item{section}{The section of the metric spreadsheet to extract data from. Can be either "off-site" or "on-site".}

\item{module}{The module of the metric spreadsheet to extract data from. Can be either "area", "hedgerow" or "watercourse".}
}
\value{
A data table including habitat_name, available_units and other relevant columns.
}
\description{
Extracts the available units for each broad area habitat, or hedgerow or watercourse type.
}
