% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_baseline.R
\name{get_baseline}
\alias{get_baseline}
\title{Extract baseline habitat details}
\usage{
get_baseline(
  filepath,
  section = c("off-site", "on-site"),
  module = c("area", "hedgerow", "watercourse")
)
}
\arguments{
\item{filepath}{The path to the metric spreadsheet.}

\item{section}{The section of the metric spreadsheet to extract data from. Can be either "off-site" or "on-site".}

\item{module}{The module of the metric spreadsheet to extract data from. Can be either "area", "hedgerow" or "watercourse".}
}
\value{
A data table including baseline_habitat_name, baseline_size, baseline_units, enhancement_size etc.
}
\description{
Extracts details from the baseline habitat parts of the metric spreadsheet.
}
