% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_creation.R
\name{get_creation}
\alias{get_creation}
\title{Extract habitat creation details}
\usage{
get_creation(
  filepath,
  section = c("off-site", "on-site"),
  module = c("area", "hedgerow", "watercourse")
)
}
\arguments{
\item{filepath}{The path to the metric spreadsheet.}

\item{section}{The section of the metric spreadsheet to extract data from. Can be either "off-site" or "on-site".}

\item{module}{The module of the metric spreadsheet to extract data from. Can be either "area", "hedgerow" or "watercourse".}
}
\value{
A data table including post_habitat_name, creation_size and post_creation_units.
}
\description{
Extracts details from the habitat creation parts of the metric spreadsheet.
}
