% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enhancement.R
\name{get_enhancement}
\alias{get_enhancement}
\title{Extract habitat enhancement details}
\usage{
get_enhancement(
  filepath,
  section = c("off-site", "on-site"),
  module = c("area", "hedgerow", "watercourse")
)
}
\arguments{
\item{filepath}{The path to the metric spreadsheet.}

\item{section}{The section of the metric spreadsheet to extract data from. Can be either "off-site" or "on-site".}

\item{module}{The module of the metric spreadsheet to extract data from. Can be either "area", "hedgerow" or "watercourse".}
}
\value{
A data table including parcel_ref, post_habitat_name, post_enhancement_units.
}
\description{
Extracts details from the habitat enhancement parts of the metric spreadsheet.
}
