% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_headline_results.R
\name{get_headline_results}
\alias{get_headline_results}
\title{Extract headline results}
\usage{
get_headline_results(filepath)
}
\arguments{
\item{filepath}{The path to the metric spreadsheet.}

\item{section}{The section of the metric spreadsheet to extract data from. Can be either "off-site" or "on-site".}

\item{module}{The module of the metric spreadsheet to extract data from. Can be either "area", "hedgerow" or "watercourse".}
}
\value{
A data table including section, module, type of figure and units.
}
\description{
Extracts headline results from the metric spreadsheet.
}
