% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vals.R
\name{get_vals}
\alias{get_vals}
\title{Get values from spreadsheet using sheet range}
\usage{
get_vals(sheet_range, n_merges = 0)
}
\arguments{
\item{sheet_range}{the range of the sheet to get values from. Must include a sheet name and a range.}

\item{n_merges}{the number of merged cells above a the sheet_range.}
}
\value{
a data frame of values from the sheet range.
}
\description{
Get values from spreadsheet using sheet range
}
