% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sankey.R
\name{make_sankey}
\alias{make_sankey}
\title{Title}
\usage{
make_sankey(
  filepath,
  section = c("off-site", "on-site"),
  module = c("area", "hedgerow", "watercourse")
)
}
\arguments{
\item{module}{}
}
\description{
Title
}
