% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_headline_results.R
\name{get_headline_results}
\alias{get_headline_results}
\title{Extract headline results}
\usage{
get_headline_results(filepath)
}
\arguments{
\item{filepath}{The path to the metric spreadsheet.}
}
\value{
A data table including section, module, type of figure and units.
}
\description{
Extract headline results
}
